Consolidate ftp from util to adapter
